Panda 0.9.7 release

+++ src/frontend_analysis/source_code_manipulation/create_address_translation.hpp
@@ -12,7 +12,7 @@
  *                       Politecnico di Milano - DEIB
  *                        System Architectures Group
  *             ***********************************************
- *              Copyright (C) 2015-2020 Politecnico di Milano
+ *              Copyright (C) 2015-2022 Politecnico di Milano
  *
  *   This file is part of the PandA framework.
  *
@@ -88,13 +88,16 @@ class CreateAddressTranslation : public ApplicationFrontendFlowStep
     * @param first_level is true if we are analyzing a parameter and not something inside a parameter
     * @param little_endianess tells if the parameter has little endianess
     */
-   void ComputeAddress(const AsnTypeRef asn_type, const unsigned int tree_parameter_type, unsigned int& bambu_address, unsigned int& taste_address, unsigned int& registers, const bool first_level, const bool little_endianess);
+   void ComputeAddress(const AsnTypeRef asn_type, const unsigned int tree_parameter_type,
+                       unsigned long long& bambu_address, unsigned long long int& taste_address,
+                       unsigned int& registers, const bool first_level, const bool little_endianess);
 
    /**
     * Return the set of analyses in relationship with this design step
     * @param relationship_type is the type of relationship to be considered
     */
-   const CustomUnorderedSet<std::pair<FrontendFlowStepType, FunctionRelationship>> ComputeFrontendRelationships(const DesignFlowStep::RelationshipType relationship_type) const override;
+   const CustomUnorderedSet<std::pair<FrontendFlowStepType, FunctionRelationship>>
+   ComputeFrontendRelationships(const DesignFlowStep::RelationshipType relationship_type) const override;
 
    /**
     * Initialize the step (i.e., like a constructor, but executed just before exec
@@ -114,7 +117,8 @@ class CreateAddressTranslation : public ApplicationFrontendFlowStep
     * @param design_flow_manager is the design flow manager
     * @param parameters is the set of input parameters
     */
-   CreateAddressTranslation(const application_managerRef AppM, const DesignFlowManagerConstRef design_flow_manager, const ParameterConstRef parameters);
+   CreateAddressTranslation(const application_managerRef AppM, const DesignFlowManagerConstRef design_flow_manager,
+                            const ParameterConstRef parameters);
 
    /**
     * Destructor
